requirements.txt
customtkinter>=5.2.0
pillow>=10.0.0
reportlab>=4.0.0
num2words>=0.5.13
python-dateutil>=2.8.2
